Renamed display/ to canvas/ and created ui/
