+++ heatmap-client/src/canvas/app.rs
