+++ heatmap-client/src/canvas/camera.rs
