+++ heatmap-client/src/canvas/canvas.rs
