+++ heatmap-client/src/canvas/geometry.rs
