+++ heatmap-client/src/canvas/input.rs
