+++ heatmap-client/src/canvas/mod.rs
