+++ heatmap-client/src/canvas/render_context.rs
