+++ heatmap-client/src/canvas/shaders/blend.wgsl
