+++ heatmap-client/src/canvas/shaders/colormap.wgsl
