+++ heatmap-client/src/canvas/state.rs
