+++ heatmap-client/src/canvas/texture.rs
