+++ heatmap-client/src/main.rs
@@ -1,8 +1,8 @@
 #![feature(let_chains)]
-use display::Canvas;
+use canvas::Canvas;
 use leptos::*;
 
-mod display;
+mod canvas;
 mod ingest;
 
 fn main() {
@@ -10,16 +10,18 @@ fn main() {
 
     let (count, set_count) = create_signal(0);
 
-    leptos::mount_to_body(move || view! { 
-        <div>
-            <button
-                on:click = move |_| { 
-                    web_sys::console::log_1(&format!("{:?}", count()).into());
-                    set_count(count() + 1)
-                }
-                style="position:absolute;top:10px;left:10px;z-index:1"
-            >"Click Me!"</button>
-            <Canvas></Canvas>
-        </div>
+    leptos::mount_to_body(move || {
+        view! {
+            <div>
+                <button
+                    on:click = move |_| {
+                        web_sys::console::log_1(&format!("{:?}", count()).into());
+                        set_count(count() + 1)
+                    }
+                    style="position:absolute;top:10px;left:10px;z-index:1"
+                >"Click Me!"</button>
+                <Canvas></Canvas>
+            </div>
+        }
     })
 }
