+++ heatmap-client/src/ui/ui_component.rs
@@ -1,0 +1,1 @@
+
